// frontend/projects/partner/src/app/api/model/bank-connection-dto.ts
/**
 * BDK-CoreFinance API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {BankDetailsDto} from './bank-details-dto';


/**
 * Bank connection information containing basic banking details
 */
export interface BankConnectionDto {
  /**
   * Partner number - foreign key reference. Optional for new bank connections (set automatically), required for updates.
   */
  partnerNumber?: number;
  /**
   * Sequential number for multiple bank connections of one partner
   */
  sequenceNumber: number;
  /**
   * Name of the bank
   */
  bankName?: string;
  /**
   * Bank code (Bankleitzahl)
   */
  bankCode?: string;
  /**
   * Bank account number
   */
  accountNumber?: string;
  /**
   * Last modification date
   */
  modificationDate?: string;
  bankDetails?: BankDetailsDto;
}